마켓 태순 커밋

+++ src/main/java/com/koreait/yougn/beans/vo/OrderVO.java
@@ -18,6 +18,6 @@ public class OrderVO {
     private String orderReceiver;
     private String orderPhoneNum;
     private String orderMemo;
-    private Long orderCount;
+    private int orderCount;
     private int status;
 }
